break down permissions for better sharing
